Move TT attention under hardware backend
